feat: cc compilation/linking as bazel subrules
This changeset is a major refactoring of the project,
aimed at brining it up to speed with the current developments of
the Bazel ecosystem.

Sailent point is the usage of Bazel subrules as a vehicle to share
the logic responsbiel for the CC code compilation, linking and
handling of the headers map provider.

The compatbility with earlier versions of Bazel is forfited.