feat(linking): explicit library-paths for transitive sols
This change ensures that during linking actions, which
concern shared library objects, all of the transitive
dependency SOLs will be explicitly added to the
library paths of the linker.

This change aims to solve the issue in which,
when toolchain does not add relative ($ORIGIN-based)
paths to RPATH (etc.), Wl,--unresolved-symbols=report-all will
result in failures, as only direct dependencies for linking
are available.