feat(hdrs_map): graceful handling of duplicates
This changeset modifies the behavior of the
logic responsible for materialization of header
files mappings - previously, if the mappings
resulted in any duplicated file entries, the
logic will error out on the built-in Bazel
ctx.actions.declare_file call.

With this changes, situations in which a duplicate
entry might appear are detected and avoided,
while user is receving a warning message. First
definiton to materialize, wins.

At the moment this behavior is enabled by default
and non-configurable.